wrestool/restable.c: Use size_t when counting array elements.
Original warning from GCC:

restable.c: In function ‘list_resources’:
restable.c:343:2: error: assuming signed overflow does not occur when simplifying conditional to constant [-Werror=strict-overflow]
  for (c = 0 ; c < rescnt ; c++) {
  ^~~

+++ wrestool/restable.c
@@ -318,7 +318,7 @@ list_ne_name_resources (WinLibrary *fi, WinResource *typeres, int *count)
 static WinResource *
 list_ne_type_resources (WinLibrary *fi, int *count)
 {
-	int c, rescnt;
+	size_t c, rescnt;
 	WinResource *wr;
 	Win16NETypeInfo *typeinfo;
